+++ lib/util/http/client/type.go
@@ -37,7 +37,7 @@ type Config struct {
 	Verb                string            `json:"verb" yaml:"verb"`
 	Headers             map[string]string `json:"headers" yaml:"headers"`
 	CopyResponseHeaders bool              `json:"copy_response_headers" yaml:"copy_response_headers"`
-	MetadataFilter      filter.Config     `json:"metadata_filter" yaml:"metadata_filter"`
+	ExtractMetadata     filter.Config     `json:"extract_metadata" yaml:"extract_metadata"`
 	RateLimit           string            `json:"rate_limit" yaml:"rate_limit"`
 	Timeout             string            `json:"timeout" yaml:"timeout"`
 	Retry               string            `json:"retry_period" yaml:"retry_period"`
@@ -61,7 +61,7 @@ func NewConfig() Config {
 			"Content-Type": "application/octet-stream",
 		},
 		CopyResponseHeaders: false,
-		MetadataFilter:      filter.NewConfig(),
+		ExtractMetadata:     filter.NewConfig(),
 		RateLimit:           "",
 		Timeout:             "5s",
 		Retry:               "1s",